Add template for generic program credential
Removes hardcoded references to XSeries and cleans up unused assets. This template can be used to render credentials for XSeries, MicroMasters, and Professional Certificate programs. Note how the program type is slugified to assist with asset retrieval. This is a trick borrowed from Django itself that makes this template significantly cleaner.

ECOM-6531