Set HTML lang tag for records pages
Make sure that the <html lang=""> tag is set for new Student Records
work.

I also cleaned up a few instances of using en-us as the fallback
language code. Our locale generated for English is 'en' not 'en-us'
and in production, we already configure 'en' as the fallback. So
devstack should be in sync too.