Add catalog app to copy Discovery data
Adds a catalog django app and accompanying 'copy_catalog' management
command. This will query Discovery and add the results to a local
copy of the catalog in the local database.

Currently, we're only interested in programs and their courses, and
only a small subset of their data. Can be expanded as needed.

LEARNER-5049