Update the Credentials program-listing endpoint to use Staff instead of (#826)
* Update the Credentials program-listing endpoint to use Staff instead of
superuser

[MICROBA-598]

The Credentials endpoint program-listing/ was meant for the support
team and internal teams to use as a debugging tool. The user type
Superuser was improperly used and now that less internal users
have that role it is unusable. Changing this to the Django Staff role
will allow internal users to access the tool again.

* Update tests and superuser access

* missed changes