ARCHBOM-1569: Add a mock toggles state endpoint
As part of Architecture's observability work, we're adding endpoints to
IDAs that can dump the current state of various toggles.  This is a mock
endpoint that accomplishes two things: 1) confirm that we can require a
staff JWT in another IDA than edxapp, and 2) allow us to test integration
of multiple IDAs into toggle state reports before the edx-toggles
djangoapp is ready for use.

This can go away once edx-toggles is ready and integrated.