Add program email configuration models
In replacement of the settings based email configuration, this allows
editing in Django Admin. It has a similar hierarchical model where the
most specific key (program, then program type, then default) that
matches becomes the setting used. In addition, this allows us to enable
or disable the sending of the emails at each level.