Add fields to Program and Organization
Adds two fields to Program and Organization that are now in the list
view for Programs in Discovery. This is a required step to move
program_details over to the in-service stored catalog.

Once released, the copy_catalog management command will need to
be ran before the followup code is deployed.

Also fixes a typo in the memcache port in devstack