Convert eslint config to non-deprecated version
eslint-config-edx is deprecated in favor of @edx/eslint-config
Most code changes were changed with eslint --fix

Changes that weren't automatic:
- Reording of methods in MasqueradeBanner.jsx. It was a noop, but it
wouldn't do it automatically.
- Removed `isPublic` and `recordDownloaded` from ProgramRecord.jsx
state. `isPublic` is a prop, not state now. `recordDownloaded` was state
that was only used for testing. Removed it and the reliant test from
ProgramRecord.test.jsx
- rewrote checkCreditPathway to use the previous state during setState
which should prevent data mismatches.