fix: fix credentials devstack provisioning after django debug toolbar change
[MICROBA-1189]
- add new Makefile target for `dev-requirements` that ensures the django debug toolbar gets installed in the docker container

What I think is happening is that (currently) when Credentials is provisioned, when installing python requirements, we run `make requirements` and `make production-requirements` back to back. A side effect here is that `pip-sync` uninstalls the django debug toolbar since it is not a requirement in the `production.txt` requirements file(?).

This will be paired with an update to the `devstack` repo `provision-credentials.sh` script that will call this new Make target to ensure all the required packages are being installed during provisioning.