MB-1167: CourseCertificateSerializer Fix (#1135)
* [fix] Add permisssions and fix CourseCertificate

The CourseCertificateSerializer needed permission for the LMS
credentials_service_worker to operate properly. I opted for the admin
user permission instead of writing its own, because the worker should be
the only one making changes with this API anyway.

I've also reworked the serializers create() function to fix it. The
update portion of update_or_create() was not properly working because of
the unique together constraint on the model. Some of the tests were
modified to reflect the change made to create().

* remove unused import

* Formatting