fix: update program credential visible date logic (#1184)
Addresses MICROBA-1303, which was made from CR-3802. A learner’s program
certificate was showing an issue date earlier than expected, earlier
than the date when they completed all the required courses.

This was because of some faulty logic in our
`_get_program_certificate_visible_date` method. It was only returning
the latest associated `certificate_available_date`. But
`certificate_available_dates` are only set on courses that are
instructor-paced and have a certificate availability date manually set.
In the case of our learner, only one course in the program had a
`certificate_available_date`--and it was one of the earlier courses
completed by the learner. The rest of the courses did not have a
`certificate_available_date` because they were self-paced. Since the
logic only returned the latest `certificate_available_date`, and because
the later courses didn’t have that particular field, the program visible
date was being returned as a date that was too early.

Now, we have updated the `_get_program_certificate_visible_date` logic
to use the course credential `created` date when there is no
`certificate_available_date`. This will ensure that the program
credential displays the date of the latest course completed, whether the
course was instructor- or self-paced.

Note: we have some questions about whether it is more appropriate to
use the course credential’s `modified` date rather than the `created`
date, but are sticking with `created` for now because that is closer to
the original design.