feat: Add dockerfile and push image to Dockerhub
This image builds on our existing (non-ansible) Dockerfiles with the
following exceptions:
- It's the first to require node to build static assets. I've followed
the pattern we use in the ansible builds where we use nodeenv and I've
also taken the node and npm versions from ansible. This pattern may
change in the future but for now I'm prioritizing consistency with the
existing images.
- This is also the first one being used in devstack, so to utilize
existing devstack code, the folder structure has to have an additional
layer. The repo folder will live at /edx/app/credentials/credentials
instead of the /edx/app/<service> that other's use.