feat: Store date_override sent from platform (#1279)
A course certificate date_override may be sent from edx-platform. This
commit handles that data on the credentials side. The date_override key
will be sent over with every post for a course certificate. If the key
has a value, save a UserCredentialDateOverride record associated with
the given UserCredential. If the date_override value is empty, delete
any UserCredentialDateOverride that might be associated with the given
UserCredential.