[feat] MB-1464: Add lms_user_id (#1312)
* [feat] MB-1464: Add lms_user_id

A small function that populates the LMS user-id when a user logs into
credentials for the first time. This will only update the lms_user_id
for users that specifically go to credentials; either to their records
page or otherwise. This will not update for users who are awarded a
prgram/course credential.