fix: Download Record/CSV creation broken in development environments (#1709)
The Credentials IDA has a feature that allows you to download a CSV representation of your public Program Records.

In our development environment (devstack) this functionality is broken because we are trying to make an analytics call to Segment with an ID retrieved from a cookie in our web request. This cookie exists in our Stage and Production environments but does not exist in development environments. This causes an exception to be thrown.

I decided on this fix over exception handling because it seems like a better backup in case this other cookie doesn't exist and will allow us to continue tracking these analytics/events via Segment in our live systems.