feat: Add ProgramCertRecord data to API response for Learner Record MFE rendering decisions.
[APER-1946]

Adds a new field to the dictionary response of the `get_record_data` function. The new data will be information about the shared ProgramCertRecord for a Program, if it exists. This data will be returned in an existing REST API to the Learner Record MFE where it will be used to conditionally render button states in an individual Program Record.

* Move private function `_get_course_data` to top of the file.
* Updated code comments and docstrings to provide additional context wherever possiible.
* Refactors the existing logic that retrieves the data presented in our Program Record page in an attempt to reduce cognitive load. I found it very difficult to follow all the logic in the original mega-function. While this is not perfect, I think it's better than it was before.
* include the UUID of a shared program record if it exists
* add tests for refactored functions and new functionality