feat: Add `lms_last_updated_at` field to the UserGrade model (#1757)
[APER-1968]

Today, we don't have a good way to understand if grade data is out of sync with the LMS. Grades are sent to Credentials via a REST API call originating from an asynchronous Celery task on the LMS side.

This PR introduces a new field to the UserGrade model named `lms_last_updated_at`. This field will store the `modified` date of the corresponding grade coming from the LMS for a learner. This way we will know exactly the last time the grade was updated in the LMS and will be able to tell how in sync the grades are from the two systems.

* Add the (optional) `lms_last_updated_at` field to the UserGrade model.
* Update existing Grades endpoint to accept this data if provided.