feat: add translations requirements
A new GH Action has been created in openedx-translations as part of the
prototype for OEP-58. The GH action generates translations. Credentials
was chosen as the first repo to be used for the OEP-58 prototype. To
streamline the process of generating translations via GH Action,
translation requirements have been created in the credentials repo.