feat: redirect users to LR MFE records pages where appropriate
[APER-1975]

It is possible for learners to generate a link to a public version of their program record(s) to allow external entities to view their progress. This is particularly important as we embed these links within messages that get shared via credit pathways. We wanted to ensure that these links are not broken when the default experience becomes the Learner Record MFE, so we will be redirecting requests to the legacy view to the Learner Record MFE.

* Implement a redirect when a public program record is being requested (and the Learner Record MFE is enabled)
* Implement a redirect when hitting the records index page (and the Learner Record MFE is enabled)
* Fix an issue with the shared program record UUID being returned with dashes (Credentials can't process UUIDs with dashes in the program record view)