fix: broken styles in program cert print media
Right now, building static assets is broken for credentials and credentials themes. This PR fixes an issues in the print media query that was blocking the build.

fix: wrong styles applied

fix: unbreak the styles that I readded in by accident

fix: we need to pin these dependencies as is, theres an issue with gulp-eslint that is causing linting errors