chore: regenerate package-lock.json after node16 upgrade
[APER-2063]

This PR is for some cleanup after the Node 16 upgrade. The aim here is to update dependencies that will cause issues and to regenerate our `package-lock.json` file in the Credentials repo.

This PR also updates the Credentials repo to use an NPM script to run our linting tests. Previously, we used `gulp` to run our linters, but `gulp` is no longer an active project (no updates since 2019) and doesn't appear to be compatible with later version of Node/NPM.

Some consequences of this PR:
1. We will stop running our (bok choy) acceptance tests suite as part of CI with these changes. They don't run successfully anymore after this upgrade and the bok choy framework has been deprecated for awhile now, there is nothing to gain by fixing them at this time.
2. We will stop running our non-React JS tests during CI because of an issue with Karma + Firefox + Node16 running in containerized GitHub action environments. I will create a followup task to explore using Chromium or another headless browser to run these tests.

Changes:
- regenerate the `package-lock.json` file
- upgrade the `eslint-config-airbnb` and `eslint-plugin-react-hooks` dependencies
- convert linting to be kicked off by NPM script, rather than using our gulpfile
- update Makefile to call NPM script (intstead of Gulp)
- update translation files
- update CI workflows to install specific versions of Node (matching what is deployed to the environment)
- don't run (bok choy) acceptance tests as part of our CI
- do not run non-React JS tests as part of CI as there is an issue with Node16 + Karma + Firefox + GitHub Actions.