fix: Allow Credentials to handle all redirects to the Learner Record MFE where appropriate (#1915)
[APER-2224]

Credentials already handles redirects for requests to the records page (credentials.blah.whatever/records) and shared program records (credentials.blah.whatever/programs/shared/{shared_program_record_UUID}/. Early in the implementation of the Learner Record MFE, we put logic in the LMS to build a URI directly to the _private_ program record pages in the LR MFE. This has proven to be a bad decision because it skips the cross-service authentication for the user.

This PR makes it so Credentials will handle _all_ redirects to the Learner Record MFE when appropriate. This solves an issue learners are experiencing when they are trying to make an authenticated REST API request before they have been authenticated with the Credentials service (APER-2224).