temp: omit publicPath from webpack.config.js
Temporary change that is a test fix for APER-2210. I read that if `publicPath` is set in a webpack config then `django-webpack-loader` will use `publicPath` over the static URL settings. I think this could be the source of the problem with some of our assets being accessed at an incorrect URL.