fix: fix lookup of `CourseCertificate` objects (#2045)
[APER-2505]

This PR fixes an issue that can come up when trying to retrieve an existing CourseCertificate object. Originally, we were using the `course_run` field to do the lookup but sometimes this data isn't set properly when it is created (which is a known issue and we're tracking this internally as APER-1777).

Instead, we should use the `course_id` field (which is really the course run key) to do the lookup.

This also fixes some errors in our unit tests that were setting the wrong data through use of our test data Factories.