fix: broken swagger and duplicate route
This API contains a non-CRUD class with no serializer, which the
swagger conversion can't handle automagically. Additionally, there was
an endpoint that had been imported twice because of a file structure
containing multiple urls.py files.

* specifies the swagger schema for a non-serializer class
* removes redunandant route via wrong endpoint

Note that the generated swagger document is somewhat confusing, and does not
correctly group the different APIs. This is unavoidable because of two
separate complicating factors:

* the openAPI specification, which cannot has certain assumptions about base prefix
* changes within this cluster of APIs, in which endpoint structure has had three
  distinct layouts, before we reached the current, edx-standard structure.
  (`appname/api/v#/endpoint_name`, `api/v#/endpoint_name`, and
  `api/appname/v#/endpoint_name`)

There are ways to address this, but they are all messy (e.g. creating aliases for old endpoints).

Fixes: APER-2531