feat: change behavior for unknown username, userID (#2058)
* feat: change behavior for unknown username, userID
If an unknown username or lms_user_id is sent, might be real in the LMS
and unknown to credentials (i.e. a student who has never received a
credential), or it might be fake. Either way, return a 200 response with
an empty status list, rather than reporting whether or not the user
exists.

Also:

* Fixed this information in both the openAPI and the ReST docs
* improved the formatting/display of the docstring in POST so it will work better in the generated swagger
* Added some Python typing
* fixed some grammar bugs in the ReST
* clarified that username/user ID is either/or (exactly one) in the ReST
* removed some commented out code coverage experiments, with the approval of @cdeery 

FIXES: APER-2425