feat: fixing an exceptional case
When this code switched from using direct ORM calls to the catalog API,
the exception handling that was there became meaningless and the error
protection was no longer valid.

This fixes the code and adds a test case. While writing test cases to
improve coverage for a few of the other exception handling cases, I
realized that those exceptions were already handled by database
constraints, so I removed the error handling code altogether.

FIXES: APER-2425