chore: pin `social-auth-app-django` to 5.2.x
[APER-2847]

The latest version (5.3.0) of `social-auth-app-django` introduces a few migrations that can cause out-of-memory problems if the Credentials `social_auth_usersocialauth` table is large.

For now, we are pinning this package to an earlier version before the problematic migrations are introduced (5.2.0).

For more info, see: https://github.com/python-social-auth/social-app-django/issues/501