feat: make catalog model fields read only
[APER-2841]

This PR updates the Catalog app's models available via Django Admin of the Credentials IDA to be read-only. We are making this change becuase Credentials is not the source of truth for this data, it is synced from the Course-Discovery IDA.

If the data needs to be updated, it should be done in Discovery and then admins should use the `copy-catalog` management command to sync the data to Credentials.

This is being done to ensure data integrity across services.