chore: status checks based on prior commit
when we upgrade codecov, it's looking at this file and enforcing 95%.
However, it's clear that historically we've been hovering around 94.69%.
I'm not sure if the old version of the action inherently had some wiggle
room. We could probably fix this by using "round: " in codecov.yml but
that might have other negative repercussions. Switching to target: auto
will guarantee the ratio never goes down.

FIXES: APER-2882