feat: add indexes to speed up django admin (#2244)
* feat: add indexes to speed up django admin

This adds indexes for the Boolean fields inherited from the abstract
models, and also for full name. This is about trying to improve the
speed of reads, but it might give us a penalty on writes, so we will
need to keep an eye on alerts from new relic.

Note that application of the migration might lock the table for a bit.
https://realpython.com/create-django-index-without-downtime/ has ways to
work around that but I don't think they are necessary. Brief locks to
credentials shouldn't be catastrophic. (When I tested it in a
low-powered development environment with about 5 million rows it took
approximately half a minute.)


FIXES: APER-2970