feat!: Remove unused feature flag
[APER-2913]

This PR removes the `USE_LEARNER_RECORD_MFE` feature flag. This was originally used to gate redirections to the Learner Record MFE when it was under development.

The Learner Record MFE is now required for program record functionality, the legacy FE was removed as part of #2104 and #2106.

Additionally, I have added a few unit tests covering redirection functionality to the Learner Record MFE.