fix: revert changes from #1912
In early October, we accepted an to the public Credentials Dockerfile presented through an OSPR. The purpose was to reduce the overall size of the Credentials Docker image.

However, shortly after this point, the Credentials image build started to fail.

In a first attempt to troubleshoot the issues, I am removing the accepted changes from PR #1912 to see if this resolves the problem. We can then look at reintroducing them if needed.