fix: display earned course credentials even if the grade is missing (#2287)
There are well documented communication issues between the LMS and Credentials systems. There are times that Credentials will just miss a grade update from the LMS. The grades stored in Credentials are used to render a learner's progress in their courses through the Learner Record UI.

This PR changes the behavior of the API that feeds data to the Learner Record MFE (program record page) and the Support Tools.

Previously, when rendering the program record page, if there are no grades available for a course (that is part of a program) we would show the course credential as "unearned".

This has caused a lot of confusion for our learners, as a missing grade in Credentials does not mean they haven't earned the certificate. In fact, most of the time, the learner has been able to view and share their course certificate and they are confused as to why it shows as "unearned" on their program record(s).

Now, if the system is aware of an earned Credential, we will populate the issue_date field (which the frontend uses to determine if a certificate is earned or not). This way we can reflect the learner record as accurately as possible for a learner.