feat:  bulk learner cert status and point (#2322)
* feat:  bulk learner cert status endpoint

Adds a bulk version of the learner_cert_status endpoint, following the pattern we use elsewhere for bulk variants: `bulk_learner_cert_status`.

Extrapolating the logic for a single learner/course combo into the API. Whether that particular structure belonged in api.py or in the specific API view is a matter of art rather than science; it felt like a more generalized function for the Python API rather than something for the view. (Reasonable developers could disagree; the specifics of the structure here are designed around the API view so you could go either way. To me it felt cleaner to leave it in `api.py`.)

This does have some repeated code as a side effect of not wanting to change the contract for the previously-existing endpoint, but it's kept to a minimum..

FIXES: APER-3102