fix: fix social media and sharing icons
The Credentials IDA gives our users the ability to share or print their program certificates if they wish. This functionality is enabled by rendering icons (provided via a third party package called `fontawesome`).

Recently, something caused this functionality to break and the icons are rendering as squares.

This PR attempts to fix this issue. Initially, we attempted to troubleshoot the original issue with the old `fontawesome` package (which was released 7 years ago). We were unable to do that. Then, we tried to update to the latest version of fontawesome via NPM. This caused additional issues that weren't well understood. Finally, I have decided to remove the npm package in favor of a pip package that is designed to work specifically with Django.

Thus, this PR adds a new default installed app to the IDA's `installed_apps` setting -- `fontawesomefree`.

This approach worked and has fixed the display of the icons used in the sharing options for a program certificate rendered by the Credentials IDA.

Additionally, an update was made to the `base.html` template to import the fontawesome icons, as well as an update to the `_actions.html` template to update the class names appropriate to the latest version of `fontawesome`.

Unrelated, this PR includes a fix for the Makefile that I ran into when trying to ensure our existing JS tests still pass after this change.