feat!: suspend building Docker image for arm64
There seems to be an issue installing a required dependency (`didkit`) for the `arm64` platform. Since one of the two platforms are failing, it causes the good image for the `linux/amd64` platform to not upload either.

This is keeping us from being able to push an updated Docker image running Python 3.11 to DockerHub and impeding Devstack developers.

We are under the impression that only 2U uses this Dockerfile and this shouldn't affect the larger community.

For additional details, please see the public Credentials issue: https://github.com/openedx/credentials/issues/2477.