feat: the signatory is not mandatory on coursecertificate configurations (#2515)
* feat: making  the django admin understand the signatory is not mandatory on coursecertificates

Credentials IDA requires signatories on `ProgramCertificate` configurations.  However, on `CourseCertificate` configurations, not all sites will care  if a signatory is present. The certificates themselves don't render in the credentials IDA, and this is cloned data from the system of record. Although this has always been nullable in the ORM because of the way django handles nullability for `ManyToManyField`,  the django admin enforces the presence of a signatory.

This keeps the  signatory required for `ProgramCertificate`, but makes it blankable for `CourseCertificate`.

FIXES: APER-3494

Co-authored-by: Jason Wesson <jsnwesson@gmail.com>