feat: attempting a new fix for the local log format problem (#2658)
The previous attempt to set a local log format failed because first the base configuration files are applied, in which the logging format is set via a function call, and then the overrides come in from edx-internal. Basically, there's a timing problem.

This is an attempt to solve that problem in the hopes that the environment variables are set before the configuration files for the application are run. This change will be made in conjunction with a PR on at ex-internal.

FIXES: APER-3805