feat: performance tuning  credentials admin
this adds a single index to  UserCredential.credential_id. You would
think this would be implicitly indexed by the  presence of a
unique_together, but it isn't. I don't want to get overly exuberant with
indices because we have a lot of costly writes  during grading already,
but I can see  using SQL `EXPLAIN`  that  making this change does
convert the django admin page from `using  filesort` to `using index`,
so theoretically this should speed it up. If it doesn't,  I will try a
different approach.