feat: adding pathway status
course-discovery is adding the concept of retired  pathways, which it
will expose with a status field through the API. We still want to ingest
retired and unpublished pathways, because we don't want to break any of
our existing foreign key relationships. However, we don't want to expose
them to either the Learner Record MFE  or the program dashboard.

Still todo:
* verification
*  documentation updates
* ADR

FIXES: APER-3929