feat: cherry pick FA icons used in Credentials UIs
Previous to this change, when collecting static assets, we would include ALL of the FA icons. This PR is an attempt to limit the icons in our bundle to only the icons used in the UI.

As part of this change, the `fontawesomefree` app is no longer needed, so it has been removed.